+++ .ci.sh
@@ -18,10 +18,8 @@ shCiPreCustom() {(set -e
     fi
     case "$GITHUB_REF_NAME" in
     alpha)
-        #
         # sync branch
         shGitCmdWithGithubToken push origin alpha:mysh -f &
-        #
         # test
         # (
         #     git push -f origin alpha:kaizhu256/betadog/alpha